Safari-611.1.10/LayoutTests/canvas/philip/tests/size.attributes.setAttribute.onlyspace.html

<!DOCTYPE html>
<title>Canvas test: size.attributes.setAttribute.onlyspace</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="50" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.setAttribute.onlyspace.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

canvas.setAttribute('width', '  ');
canvas.setAttribute('height', '  ');
_assertSame(canvas.width, 300, "canvas.width", "300");
_assertSame(canvas.height, 150, "canvas.height", "150");


});
</script>
