LayoutTests/canvas/philip/tests/size.attributes.setAttribute.onlyspace-expected.txt

< [index] >
size.attributes.setAttribute.onlyspace
Parsing of non-negative integers in setAttribute
References: size.nonnegativeinteger size.error
Actual output:
Expected output:

Passed
